rename entry file
