+++ package.json
@@ -4,7 +4,7 @@
   "description": "A little script to find duplicate files by content in a given directory.",
   "main": "index.js",
   "scripts": {
-    "start": "ts-node src/find-duplicate-files.ts"
+    "start": "ts-node src/index.ts"
   },
   "repository": {
     "type": "git",
